contrib/pagestore: harden delta read plan (review on #11)
Codex review on the delta-layer / read-plan PR:

- ps_read_plan_build: image-layer lookup returning -1 (unreadable / bad footer)
  was treated as a miss, so the plan could pick an older base (or none) and still
  succeed -- the materializer would then apply deltas to the wrong base.  Now
  propagate the read/format error.
- ps_delta_layer_collect: stopping at the output cap and returning 0 silently
  truncated the redo chain (a page with > PS_PLAN_MAX_CHAIN deltas in one layer
  would be reconstructed from a partial chain).  Return -1 on overflow instead.
- PsDeltaOut.data_off was 32-bit while the on-disk offset and the layer-store read
  API are 64-bit; a payload past 4 GiB would be read from a wrapped offset.  Widen
  to uint64_t and drop the narrowing cast.

Deferred (same as image #3): verifying foot.data_crc needs a per-payload checksum
(format bump); index_crc is verified.

Layer unit test grows an overflow case -> 27/0; standalone 119/0.

+++ contrib/pagestore/pagestore_layer.h
@@ -206,7 +206,7 @@ typedef struct PsDeltaRec
 typedef struct PsDeltaOut
 {
 	uint64_t	lsn;
-	uint32_t	data_off;
+	uint64_t	data_off;
 	uint32_t	data_len;
 } PsDeltaOut;
